[user] clean: aur moved to user
The key idea here is that all packages that people will request will be put in the user.conf

+++ 02_userspace_install.sh
@@ -1,19 +1,24 @@
 #!/bin/bash
 
 # Download yay and install with makepkg
-echo "==========================================================="
-echo "## Cloning and bilding yay (requires sudo)"
-echo "==========================================================="
-git -C ~ clone https://aur.archlinux.org/yay.git
-(cd ~/yay && makepkg -si)
-# Tidy up
-rm -rf ~/yay
+if ! command -v yay &> /dev/null
+then
+
+    echo "==========================================================="
+    echo "## Cloning and building yay (requires sudo)"
+    echo "==========================================================="
+    git -C ~ clone https://aur.archlinux.org/yay.git
+    (cd ~/yay && makepkg -si)
+
+    # Tidy up
+    rm -rf ~/yay
+fi
 
 # Install packages listed in packages/aur.conf
 echo "==========================================================="
 echo "## Install packages from the AUR"
 echo "==========================================================="
-yay -S $(grep -v "^\(^#.*\|[ \t]*\)$" packages/aur.conf | sed 's/[ ]*#.*//g' | tr '\n' ' ')
+yay -S $(grep -v "^\(^#.*\|[ \t]*\)$" packages/user.conf | sed 's/[ ]*#.*//g' | tr '\n' ' ')
 
 echo "==========================================================="
 echo "## DONE :)"
